requirements.txt
Flask==1.0.2
PyYAML==3.12
